+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.3.0-rc.25",
-  genVersion: "2.803.3",
-  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.25 2.803.3 2.0 @orq-ai/node",
+  sdkVersion: "4.3.0-rc.26",
+  genVersion: "2.806.4",
+  userAgent: "speakeasy-sdk/typescript 4.3.0-rc.26 2.806.4 2.0 @orq-ai/node",
 } as const;