+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-04T13:35:53.114Z",
+    "2026-02-05T05:06:29.971Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {