+++ packages/orq-rc/RELEASES.md
@@ -6409,3 +6409,13 @@ Based on:
 - [typescript v4.3.0-rc.25] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.25] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.25 - packages/orq-rc
+
+## 2026-02-05 05:09:22
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.704.0 (2.806.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.26] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.26] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.26 - packages/orq-rc